requirements.txt
numpy>=1.26
pandas>=2.1
scikit-learn>=1.3
joblib>=1.3
typer>=0.9
rich>=13.6
matplotlib>=3.8
pywebview>=4.4
